DX: adapt Prelude workaround and cleanup internal

+++ scripts/clean-dist.sh
@@ -31,7 +31,7 @@ do
   src_f2="${src_f}x"
   raw="${f%.js}"
   if [ ! -f "$src_f" ]; then
-    if [ ! -f "$src_f2" ]; then
+    if [ ! -f "$src_f2" && $raw !== "Prelude" ]; then
       echo "Removing $raw.js"
       rm -f $raw.js $raw.js.map $raw.d.ts $raw.d.ts.map
     fi
